move files; change paths
