+++ tests/unit/modules/contact/ContactTestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace hiapi\heppy\tests\unit\modules\contact_module;
+namespace hiapi\heppy\tests\unit\modules\contact;
 
 use hiapi\heppy\tests\unit\TestCase;
