+++ tests/unit/modules/contact/ContactUpdateTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace hiapi\heppy\tests\unit\modules\contact_module;
+namespace hiapi\heppy\tests\unit\modules\contact;
 
 class ContactUpdateTest extends ContactTestCase
 {
