+++ tests/unit/modules/domain/DomainInfoTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace hiapi\heppy\tests\unit\modules\domain_module;
+namespace hiapi\heppy\tests\unit\modules\domain;
 
 
 use hiapi\heppy\tests\unit\TestCase;
