+++ tests/unit/modules/host/HostCreateTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace hiapi\heppy\tests\unit\modules\host_module;
+namespace hiapi\heppy\tests\unit\modules\host;
 
 use hiapi\heppy\tests\unit\TestCase;
